Fix unexpected CoreException
  "org.eclipse.core.runtime.CoreException: No property tester contributes
  a property projectPersistentProperty org.eclipse.team.core.repository to
  type class org.eclipse.core.internal.resources.Project"

which was logged whenever the context menu was opened on a project.

The cause of the Exception was that the value of the "property"
attribute of the test element was wrong. The method "testProperty"
of org.eclipse.core.internal.propertytester.ResourcePropertyTester
takes both the property name and the property value from the "args"
argument.  The value of the attribute "property" is used to determine
the type of the property.

The EGit UI behaviour is the same as before, but without the
exception in the Error log and with a cleaner plugin.xml.

Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>