Remove legacy tag_egit.sh
This script should have been removed when we imported the project
history to Eclipse.  It was previously used when we had a single
unified repository with JGit, and we were building less standard
releases hosted on the jgit.org domain.  Few people, if anyone,
beyond me even understood it.  So lets just get rid of it.

Change-Id: I8087c76ca97cf10a251b1e3ad9e549c20ea13036
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>