Make tracing 3.4 compatible and plug-in local
Sharing trace classes between plug-ins is bad, as the OSGI tracing
infrastructure will write the ID of plug-in that exposes the trace
class into the trace file. Instead the ID of the plug-in that is
writing the trace should be written. This is achieved by obtaining a
separate Location for each plug-in.

In order to achieve 3.4. compatibility, a temporary trace
implementation using System.out is used.

Bug: 306793

Change-Id: 	I2765961f9c0b886995e74b03404e1f20aadf7b28
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>