BranchSelectionDialog fixes
Currently, hitting "Rename" on a tag
will do some renaming inappropriately: the result
is a new local branch.
The same goes when renaming a remote
branch using this dialog: the result is a local
branch.
This fix will disable renaming of tags
altogether and use the appropriate prefixes
when renaming branches.
In addition, the different buttons will be
enabled/disabled depending on the selection
of the tag/branch.

Change-Id: I38b2f04f71ba7784506891fb22bc2e72bfaec4ec
Bug: 308457
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>