Fix EGit history handling code to consider IAdaptables
The history code naively assumes that all of its inputs are of type
IResource. However, this is invalid as objects like an editor's
input or JDT's handling of the model of a Java project are not
represented by an IResource. The history code needs to be corrected
to consider the case where an IAdaptable is handed to it.

Bug: 311249
Change-Id: Ie1e20f68b9023bd98981f42d9b8fa14aae250614