EGit operation / action refactoring
Replace IWorkspaceRunnable in EGit operations (e.g. BranchOperation)
by a new interface IEGitOperation.
IWorkspaceRunnable is no good interface for an operation. The user 
thinks he has to execute the operation using Workspace.run. But that 
is not right for all EGit operations. Currently many operations are 
used by just calling op.run. Thus a new interface with an execute
method was introduced. Those operations that need a workspace runnable
and a scheduling rule do the workspace runnable in their execute 
method (e.g. branch). It's an internal of an operation how it needs
to execute.
BranchOperation now uses a workspace runnable to lock the workspace
while the working tree is modified.
BranchAction executes BranchOperation in a Job. The other actions
will be adopted to this execution pattern in further commits.

Change-Id: Ia3c1bf157d17632a17b7c678ee45717fb7944f0c
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>