Fix RepositoriesViewLabelProvider
If there are two branches with the same shortened name (e. g. a local 
and a remote "master" branch), the current implementation wrongly 
marks both of them as "checked out" if one of them is checked out.

The fix looks for the full branch name instead of the shortened name
in order to determine whether the icon should be decorated.

Change-Id: 	Ib6652732cfecfa762c0aa810cb4a5226b46fc27b
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>