Fix failing SharingWizardTest.shareProjectWithNewlyCreatedRepo
This test failed when run from maven because the page where the team
provider selection would occur was short circuited by Eclipse, since
only one provider exists. We could skip that part and the test would
run in maven, but not using the default options from Eclipse, or as
this patch does, add another team provider, e.g. CVS to the running
test.

Bug: 314571
Change-Id: Ia88953e6d4d4160aede97b5f0c76c9cccc08a839
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>