More strict rules for guessing clone URL
URL guessing was unreliable. E.g. simple file system paths in many
cases are not pointing to a local git repository but were accepted
which irritates users. If the string in the clipboard ends with .git or
is starting with git or git+ssh uri scheme the chance to generate
false positives is very low.

Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>