Refactor the BranchSelectionDialog
This refactors the generic part of the BranchSelectionDialog
into an abstract class for better re-usability. The concrete
subclasses have been adjusted accordingly.
The specific reason for this is that the MergeTargetSelectionDialog
must deactivate the "OK"-button when the currently checked-out
branch is selected.
The texts used in the different classes were refactored so that
their keys correspond to the repective classes they are used in.

Change-Id: Ie70dfbd3eb7564032db1c0be86bf5ae7473d1ced
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>