Fixed deprecated API usage
Besides fixing the deprecated API call, also moved the sleep in AppOperationTest
to a more reasonable location. For me, the tests failed most of the time as localTime
and the timestamp in the repo were the same. I think it makes more sense to wait
after the commit to have a time gap between commit and file change

Change-Id: I3c481096f1b89f35b0b232fbcd8f4b1ad0d012fc
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>