Launch properly configured Compare view from Workspace model
In GitModelSynchronizeParticipant.asCompareInput() we are able to detect
that user launch Compare view form Workspace model. In this case
instance of IFile will be used in asCompareInput parameter. Base on this
knowledge we can create proper configuration for Compare view based on
file location and data in GitSynchronizeData.

Since now we can use also Workspace model presentation when user launch
synchronization that don't include locally made changes. In this case we
are able to show user proper data in Compare view.

Bug: 324826
Change-Id: I179c1090a896f246bedb7c84399f18488d359c9a
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>