Add check boxes to sharing wizard
In the sharing wizard only the projects were shared which were
selected in the tree. This was not clear for the user. A user always had
to actively select the entries before he could press finish, although he
already had selected them before he started the wizard. Furthermore
the selection got lost when the user created a new repository.

Now the check boxes are relevant for sharing.

Bug: 307112
Change-Id: Idc77dd58c9630aa10683d1d0156a95a4bd44e175
Signed-off-by: Stefan Lay <stefan.lay@sap.com>