Integration of a merge tool
This adds a "Merge Tool" action to the team menu which is only
active if the repository of the selected resources is in state
"Merging".

When selected, a dialog will appear asking whether the workspace
(pre-merged) or HEAD (un-merged) version of the conflicting files
should be used as input for the merge editor's left pane.

The selected files will be shown in the upper part of the editor
(only the files which have conflicts or whose content differs
between the workspace and HEAD version). On double-click, the actual
merge editor is opened. The user can now edit the left part and save
the changed content to the workspace.

Clicking next/previous the user can navigate to the next/previous
conflict/change (even across file boundaries).

The choice on the first dialog is only relevant for files with
conflicting changes, i.e. changes that cannot be auto-merged.
For files which can be auto-merged without conflict, always the
workspace content will be shown on the left side.

The right side either shows the version found at MERGE_HEAD (in
case of file with conflicts) or HEAD (in case of file which could
be auto-merged without conflicts).

Change-Id: I6d931c24a7122466b566fa61104ea2dbdf72356e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>