Interactive commit should delete empty trees
The commit operation as used from the commit dialog did not delete
trees when they became empty. Today Git does not leave empty trees,
it is not intended here and the user certainly does not expect it.

We also add a testcase for this, though the tests for the CommitOperation
lack a lot in general as the logic is not tested anywhere else. Moving
to DirCache will probably reduce the number of needed tests at this
place.

Bug: 314105

Change-Id: I897e364e476f942f8c5361a809f5fd849a972fee
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>