Fix the org.eclipse.egit.ui.test Tycho build
Change I874fc385dca9559f892e850d2a95e67bf51c5ccf tried to import
the org.hamcrest package instead of a bundle dependency to
org.hamcrest. This works in Eclipse but unfortunately not in the Tycho
build. Hence reverting this change until a solution for this problem
(bug 334540) is found.

Change-Id: I9a0e58c495b426612b8db9ecd5fba34caa86b0f0
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>