Add a "Fetch Change from Gerrit" action
The action is available from the Repository view (on Repository nodes)
as well as via Team > Remote > Fetch a change from Gerrit.

The user needs to select a URI (the dropdown is populated from all
URIs available from all remotes of the current Repository) and a
change. There is a content assist for the change. By clicking
CTRL+Space, the user can see a list of changes and select one.

The user can also paste the change in the form

refs/changes/<nn>/<nnnn>/<nn> from the Gerrit WebUI or type it in
manually if they know the specifics...

There are several options available to perform after fetch was done:
creation of a local branch (default), creation of a tag, checking
out FETCH_HEAD, or doing nothing else (just updating FETCH_HEAD).

If a branch or tag is to be created, a name is suggested in some
pseudo-hierarchical format ("change/<nnnn>/<nn>" which comes in handy
when the user activates the "hierarchical branch layout" in the
Repositories view)...

Change-Id: Ie639e3662cc3994331e26c8c94e11458ad9dd8fa
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>