Remove redundant command definitions
This tries to reduce the number of duplicate command definitions in
plugin.xml which were done historically in order to achieve
different behavior in different environments, e.g. "Commit" in
the team menu would behave differently from "Commit" in the
Repositories View context menu.

Instead of having two commands for "Commit", we should have one
command and two different handlers with mutually exclusive
"activeWhen" conditions that handle this command.

The advantage is that we have only one key binding and command name
and that the plugin.xml becomes easier to maintain.

Change-Id: I1082ad095b20e2e05536fc10f6b521a151344e2e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>