Improve exception handling during resource decoration
In case a resource cannot be decorated, do not show/ log an error, but
silently ignore this problem (and trace the information). Also keep
track of 'not decoratable' resources and do not try to decorate them
again.

Catch all exceptions that occur during decoration and trace them. Do
not show/ log them as errors - they are only confusing.

Bug: 337116
Bug: 337671
Change-Id: Iae63ecf51b47073c9216762b3d2dee8c458dbb77
Signed-off-by: Philipp Thun <philipp.thun@sap.com>