Reduce number of LabelProviderChangedEvents
When a resource changes there are several events fired that lead
towards re-decoration of this resource (and parent resources). These
events are often redundant, e.g. when checking out a different
branch, JGit fires a RefsChangedEvent and an IndexChangedEvent.
Furthermore Eclipse detects that a resource changed and thus fires a
ResourceChangeEvent. But in the end only one re-decoration should be
triggered, thus a mechanism is introduced that delays firing an event
for 100ms and ignores further events within this time frame.

Change-Id: I4dfb4156e289d573f5f115613f3c55604e9c6746
Signed-off-by: Philipp Thun <philipp.thun@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>