Reduce screen blinking before showing Fetch wizard
Because lack of support for check box column in tables in SWT we are
using a hack in Fetch wizard to obtain image for four check box states
(enabled-checked/unchecked and disabled-checked/unchecked).

Previous implementation was creating four times a small window with
one of given check box state. Now we are including all states in
single window, then we are creating proper representation based on
part of created window.

This patch also improve overall Fetch wizard creation time.

Change-Id: I7368eca8c85b2ad128089229c6099f347f1e632f
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>