Revert "Use CompareUtils.compareHeadWithWorkspace"
This reverts commit 19f2cf44d7868aa2d4b5faa811af5dbbc9600a12. This
commit changed the behavior so that it opens a new top level window when
you select the 'Compare with->HEAD Revision' menu.  It used to create a
new editor.  This change in behavior is a regression since the top-level
window is disruptive to a standard multi editor workspace.  It also
means standard functions like search and editing are no longer
available. Reverting this commit fixes this regression.

Change-Id: I16ceaa0f36c54aad63b87471d4a5145c7efe0da3
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>