Teach the Sharing wizard to share with an existing Repository
This uses the Refactoring capabilities available in Eclipse to allow
sharing an Eclipse project with a Repository that is not located in a
parent folder of the project.

By default, the Sharing Wizard now allows selection of an existing
Repository (from the RepoView) and optionally a relative path
within this repository into which the project is to be shared (always
using the project name as terminating folder).

Upon finish, the project location is then moved to the new target folder
(a child folder of the working directory of the selected
Repository) before the ConnectProviderOperation is executed.

Users can click a checkbox to enable the previous behavior where the
Repository can only be created in a parent folder of the selected
project or projects.

Bug: 339486
Change-Id: I5e0f34e331b5427801f06ad8026ff1a78b488333
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>