Fix ClassCastException in LinkWithSelectionCommand
Use command framework to manage state of the toggle menu item [1].
Reacting to the state of an SWT item in a command handler is a bad idea
since if the command would be ever shown in 2 places (a toolbar and a
menu, for example) the state will get out of sync. 

In addition this patch ensures that the toggle state is persisted across
platform restarts.

[1] http://blog.eclipse-tips.com/2009/03/commands-part-6-toggle-radio-menu.html

Bug: 348780
Change-Id: Ie0a6848c52c27dd843d388ad0fe3ba1ab191978c
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>