[repoView] Open files in content-type specific editor.
Updates the "Open in Text Editor" command to be "Open in Editor".
This affects the double-click action as well as context menu action
in the Git Repositories view to no longer be hard-wired to open
selected files in the default text editor.

The command now attempts to resolve the selected path to an IFile
handle and will open an editor on that IFile if it can be located.

Bug: 355806
Change-Id: Iebca7380fd3a8ab2908c50eb831cd2a7120e606d
Signed-off-by: Kevin Sawicki <kevin@github.com>