Perform project creation within a workspace operation.
Originally, when using a GIT Team Project Set to setup a new workspace
(which would eventually contain many projects) if auto-build was
enabled, Eclipse would start building too early when the workspace is
not yet complete, leading to annoying build errors. Now the execution
within a workspace operation ensures that the build is
not started before all projects are created/imported.

Bug: 358530
Change-Id: Ic69e9cf01bf20dc567590370b5cbfc1ee3030631
Signed-off-by: Renato Bertacco <renato.bertacco@finantix.com>