[stagingView] Ensure reload occurs on UI-thread
This adds more guards and checks to the process of reloading
the staging view with a specified repository.  Reload clears
the current UI if a null repository is specified and so the
method should now always run on the UI-thread.  Also more strict
checks should be added to ensure the repository being displayed
is valid and registered.

Change-Id: I6f8cfca34c1baa4085b9c2f1e929d009a0cb0af0
Signed-off-by: Kevin Sawicki <kevin@github.com>