Suggest name when source ref is selected in new branch dialog
When using Team > Switch To > New Branch..., the user can select a
remote-tracking branch to create a local branch based on it. Before this
change, the user had to manually enter the name of the branch (which is
error-prone). With this change, the name is automatically suggested
based on the remote name.

So when "refs/remotes/origin/stable-1.0" is selected, we pre-fill the
text with a suggestion of "stable-1.0", which is usually what is wanted
anyway.

We keep updating the suggested branch name in case the user chooses a
different source ref, but stop doing that when the name is changed
manually.

Change-Id: I061d33dad289d93d6be41e1e902ed373168b1e2c
Signed-off-by: Robin Stocker <robin@nibor.org>