Add support for Eclipse-SourceReferences header
Add extensions required to import a plug-in using SCM URL from its
MANIFEST.MF. To give it a try, open Plug-ins view, select a plug-in,
right click and Import As > Project from a Repository.... Then follow
instructions in the wizard. Importing with a tag is not supported yet,
see bug 367712.

This allows EGit compilation against all supported eclipse platform
versions but only builds the p2 repository when built against Juno.
This should be improved in a subsequent change.

Bug: 327381
Change-Id: Ia4eb34fec24f5fe714a1b5a2b50fe0393e8d3f39
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>