[historyView] Load commit body lazily
The body of loaded objects is now thrown away in the walker after it has
parsed the information necessary to render the history graph. It is
reloaded when content of the body is needed in the HistoryView.

For large repositories this saves a lot of memory, e.g. about 110 MB for
linux kernel.

JGit-Dependency: I9291ba8c34c8744bc009b1bd302ed28bfa4e9476
Change-Id: I1504a6127e7fc2ebea67baade5e9799b256ab116