[repoView] Show name of checked out tag in repository label
Make the project decoration and repository decoration be
consistent when showing a non-symbolic ref.

This moves the logic to compute a short branch string
for a repository to RepositoryUtil to be accessed by
both the decorator for resources and the decorator
for repository view elements.

Updates the following decoration styling:
  * Ref name is now shortened so "refs/tags" prefix is dropped
  * Ref name is shown before abbreviated SHA-1
  * Parentheses have been removed
  * Ellipsis has been removed
  * Single space between ref name and SHA-1

Change-Id: Id8a1b6089763fd17980de7640fd42f14f2d41c2e
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>