[historyView] Use single resource manager
GitHistoryPage now manages a single ResourceManager
instance that is passed to all downstream objects to
be used for creating colors.

This removes the need for downstream objects to worry
about managing and disposing of colors and instead
ensures that all colors are properly disposed of only
when the history view itself is disposed.

Change-Id: Ifbf35e92c2d930c4e448f3cefe7d907c4e0d9bad