[stagingView] Add Delete action for untracked files
If the selection only includes workspace resources, the normal
DeleteResourceAction is used. Because of this, the delete can even be
undone. This change also makes the undo and redo menu items available
when the staging view is focused for this to be discoverable.

If the selection includes a non-workspace resource, the action is
carried out by a new DeleteResourcesOperation, which can handle
non-workspace resources. This has no undo support though.

The new operation can later also be used for offering Delete for Working
Directory files in the Repositories view.

Bug: 384609
Change-Id: I38f68d19877a2be495fbc5da30b17503fbf0dd86
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>