[stagingView] Allow to commit using Ctrl+Enter (Cmd+Enter)
Otherwise mousing to the small toolbar icon is always needed.

The shortcut is Mod1+Enter, so it's Ctrl+Enter on Linux and Windows, and
Cmd+Enter on OS X, matching other applications providing this shortcut.

The shortcut is shared with CommitDialog, which ensures consistency.

To make it discoverable, it's shown in the tooltip of the toolbar
action.

Bug: 382936
Change-Id: I24f313543c1bd27c8985b332b2c6d07b20ed5733
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>