[historyView] Fix Create Patch result diff when filtered
When the History view is filtered to only show commits of the selected
resource, creating a patch resulted in an incorrect diff. The problem
was that the RevCommit was used as is, and the parent was the one as
shown in the History view, not the real parent of the commit.

Test with e.g. ApplyCommand from JGit, filter the History view to
resource, and then Create Patch. The diff should be pretty small.

Change-Id: I2afa11033b26069795f57d2381e66973865cd7a9