Restore previous commit message when amend is deselected
Sometimes the user clicks on amend by accident, which would cause the
carefully crafted commit message to be replaced by the amended message.

Now it is saved when selecting amend and restored when deselecting
again.

The pre-amend message is reset when the commit message component is
reconfigured, so it won't survive a repository switch in the staging
view. But for the accidental click it's enough.

Bug: 364022
Change-Id: Icdf56dec24459aa66cc9bf8ab70aa957de39c31d